Renamed lib to shared, renamed from util to utils
