+++ etl/main.py
@@ -2,7 +2,7 @@ import os
 
 from data_enhancement.enhance_data import run as run_enhancement
 from data_extraction.scrape_data import run as run_extraction
-from lib.util import write_data_to_json, read_data_from_json
+from shared.utils import write_data_to_json, read_data_from_json
 from upload_to_elasticsearch.elastic import write_to_elastic
 
 SRC_PATH = os.path.dirname(os.path.dirname(os.path.abspath(__file__))) + '/src'
